fix: cast cache result to int in getStarsCount
Cache::remember returns mixed, causing type error when cached
value is stored as string

+++ app/Services/GitHubService.php
@@ -24,7 +24,7 @@ final readonly class GitHubService
     {
         $cacheKey = "github_stars_{$owner}_{$repo}";
 
-        return Cache::remember($cacheKey, now()->addMinutes($cacheMinutes), function () use ($owner, $repo): int {
+        return (int) Cache::remember($cacheKey, now()->addMinutes($cacheMinutes), function () use ($owner, $repo): int {
             try {
                 /** @var Response $response */
                 $response = Http::withHeaders([
